Revert "MB-100 indexer: Account for memdb jemalloc allocated memory"
This reverts commit b60ceb0aee0389587b57bdffc069909ea64c3eef.

+++ secondary/memdb/mm/malloc.go
@@ -59,6 +59,6 @@ func Stats() string {
 	return s
 }
 
-func Size() uint64 {
-	return uint64(C.mm_size())
+func Size() int {
+	return int(C.mm_size())
 }
